Merge branch antcode_repo/feat/bcdns_support of git@code.alipay.com:antchain/open-relayer.git into develop
https://code.alipay.com/antchain/open-relayer/pull_requests/7

+++ r-dal/src/main/java/com/alipay/antchain/bridge/relayer/dal/entities/SDPMsgPoolEntity.java
@@ -27,7 +27,7 @@ import lombok.Setter;
 @TableName("sdp_msg_pool")
 public class SDPMsgPoolEntity extends BaseEntity {
     @TableField("auth_msg_id")
-    private Integer authMsgId;
+    private Long authMsgId;
 
     @TableField("version")
     private Integer version;
